Adds Color_orange and a corresponding color tag to colors.tbl.
Inserts "Orange" between "Bright Yellow" and "UI Light Green" and adds the default color tag "$o" to use it. (Prerequisite for using orange as a ping display color.)

Also added some comments to identify the RGB values by the corresponding color name, because as it turns out, they're awfully hard to read as-is.